Unite the certificate-related classes in one file
