+++ locust/src/ejbca/certificate.py
@@ -1,5 +1,5 @@
 """
-EJBCA certificate creation.
+Certificate-related classes.
 """
 import logging
 import json
@@ -7,9 +7,11 @@ import requests
 from OpenSSL import crypto
 
 from src.config import CONFIG
+from src.utils import Utils
 
 LOGGER = logging.getLogger('urllib3')
 LOGGER.setLevel(logging.CRITICAL)
+
 
 class Certificate:
     """
@@ -17,6 +19,8 @@ class Certificate:
     """
 
     def __init__(self, thing_id):
+        Utils.validate_thing_id(thing_id)
+
         self.c_name = thing_id
         self.key = {"raw": "", "pem": self.generate_private_key()}
         self.csr = {"raw": "", "pem": self.generate_csr()}
@@ -95,12 +99,10 @@ class Certificate:
         return crypto.dump_certificate_request(crypto.FILETYPE_PEM, req).decode("ascii")
 
 
-    def create_ejbca_user(self, _certificate_profile_name="CFREE",
-                          _end_entity_profile_name="EMPTY_CFREE") -> None:
+    def create_ejbca_user(self) -> None:
         """
         Makes a requisition to EJBCA to create a user.
         """
-        ca_name = CONFIG['security']['ejbca_ca_name']
         # create the ejbca user
         req = json.dumps({
             "username": self.c_name
@@ -152,3 +154,129 @@ class Certificate:
         """
         self.sign_cert()
         self.crt["pem"] = self.save_crt()
+
+
+class Thing():
+    """
+    Device certification data high-level class.
+    """
+    def __init__(self, tenant: str, device_id: str):
+        thing_id = Utils.create_thing_id(tenant, device_id)
+
+        self.tenant = tenant
+        self.device_id = device_id
+        self.cert = Certificate(thing_id)
+        self.thing_id = thing_id
+        self.private_key = self.cert.key["pem"]
+        self.thing_certificate = self.cert.crt["pem"]
+
+    def renew_cert(self) -> None:
+        """
+        Renew a certificate.
+        """
+        self.cert.renew_cert()
+
+    def get_args_in_dict(self) -> dict:
+        """
+        Returns the ID, private key and certificate in a dict.
+        """
+        return {
+            "thing_id": self.thing_id,
+            "private_key": self.private_key,
+            "thing_certificate": self.thing_certificate
+        }
+
+
+class CertClient:
+    """
+    Handles certificate-related operations.
+    """
+
+    @staticmethod
+    def get_private_key_file(device_id: str) -> str:
+        """
+        Creates the key filename.
+        """
+        Utils.validate_device_id(device_id)
+        return "{0}.key".format(device_id)
+
+    @staticmethod
+    def get_certificate_file(device_id: str) -> str:
+        """
+        Creates the certificate filename.
+        """
+        Utils.validate_device_id(device_id)
+        return "{0}.crt".format(device_id)
+
+    @staticmethod
+    def create_cert_files(thing: Thing, directory: str = "/cert/") -> None:
+        """Creates the .key and .crt files for a device.
+
+        Args:
+            device_id: device's identification.
+            thing: Thing object with certificate's info.
+            directory: directory to save the files.
+        """
+        thing_path = directory + thing.device_id
+
+        try:
+            with open(thing_path + ".key", "w") as key_file:
+                key_file.write(str(thing.private_key))
+
+            with open(thing_path + ".crt", "w") as key_file:
+                key_file.write(str(thing.thing_certificate))
+
+        except Exception as exception:
+            logging.error("Error: %s", str(exception))
+
+    @staticmethod
+    def new_cert(tenant: str, device_id: str) -> Thing:
+        """
+        Creates/renovates the certificate for a device.
+        """
+        Utils.validate_tenant(tenant)
+        Utils.validate_device_id(device_id)
+        thing = Thing(tenant, device_id)
+
+        return thing
+
+    @staticmethod
+    def revoke_cert(thing: Thing) -> None:
+        """
+        Revokes a certificate for a specific device.
+        """
+        # Loads the certificate as a X509 object
+        cert: crypto.X509 = crypto.load_certificate(
+            crypto.FILETYPE_PEM, thing.thing_certificate)
+        # Retrieves the Serial Number in Hexadecimal
+        serial_number = hex(cert.get_serial_number())[2:]
+        # URL to revoke a certificate
+        url = CONFIG["security"]["ejbca_url"] + \
+            "/ca/CN={0},O=EJBCA/certificate/{1}".format(
+                CONFIG["security"]["ejbca_ca_name"], serial_number)
+
+        requests.delete(url)
+
+    @staticmethod
+    def has_been_revoked(thing: Thing) -> bool:
+        """
+        Verifies whether the certificate has been revoked or not.
+        """
+        # Loads the certificate as a X509 object
+        cert: crypto.X509 = crypto.load_certificate(
+            crypto.FILETYPE_PEM, thing.thing_certificate)
+        # Retrieves the Serial Number in Hexadecimal
+        serial_number = hex(cert.get_serial_number())[2:]
+        # URL to verify the certificate status
+        url = CONFIG["security"]["ejbca_url"] + \
+            "/ca/CN={0},O=EJBCA/certificate/{1}/status".format(
+                CONFIG["security"]["ejbca_ca_name"], serial_number)
+
+        res = requests.get(url)
+        res_json = res.json()
+
+        if res_json.get("status") and res_json["status"].get("return"):
+            return res_json["status"]["return"]["reason"] == 0
+        else:
+            logging.error("Error: invalid response from EJBCA")
+            return False
